# more efficient getindex methods for unit triangular types