@tkelman 2015-12-28
The error on appveyor looks real, an UndefRefError in linalg/triangular? https://ci.appveyor.com/project/StefanKarpinski/julia/build/1.0.12133/job/v9kpyew4v3g1fmr0