@andreasnoack 2015-12-28
Great. Thanks for the detailed timings. My rule of thumb is that a memory load is more expensive than the branch, but I wasn't aware that the memory load in `zero(A.data[i,j])` was optimized away.

I'm not sure what is causing the error on AppVeyor.