@Sacha0 2015-12-28
> I wasn't aware that the memory load in zero(A.data[i,j]) was optimized away.

That optimization surprised me --- pretty clever!

> I'm not sure what is causing the error on AppVeyor.

Neither I. Everything passes locally and the offending code (b1a2d54d0b29824578e9e05ff5133f210ae6e464/test/linalg/triangular.jl#L155) looks innocuous. Is debug-play in the build environment possible, without repeatedly pushing new branches? Does MPFR behave differently on Windows?