@tkelman 2015-12-28
There's a way you can access build workers remotely over rdp, see http://www.appveyor.com/docs/how-to/rdp-to-build-worker - and everything should run equivalently if you turn appveyor on for your fork, may have to temporarily add branch names to the whitelist in `appveyor.yml` if you do it that way though.

> Does MPFR behave differently on Windows?

Possibly. C long is 32 bits on Win64, unlike most other 64 bit systems.

Also note your links to specific lines of code aren't coming out quite right - if you paste the entire url, as in https://github.com/JuliaLang/julia/blob/b1a2d54d0b29824578e9e05ff5133f210ae6e464/test/linalg/triangular.jl#L155, github will hide the `https://github.com/JuliaLang/julia/blob/` part automatically.