@Sacha0 2015-12-28
> There's a way you can access build workers remotely over rdp, see http://www.appveyor.com/docs/how-to/rdp-to-build-worker - and everything should run equivalently if you turn appveyor on for your fork, may have to temporarily add branch names to the whitelist in appveyor.yml if you do it that way though.

Great, thanks! I'll check that out if the potential issue with this PR described in https://github.com/JuliaLang/julia/issues/14490#issuecomment-167652869 gets resolved.